% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-nhanes.R
\name{load_nhanes}
\alias{load_nhanes}
\title{Load the nhanes data}
\usage{
load_nhanes()
}
\value{
A data frame containing the nhanes dataset.
}
\description{
Load the nhanes data
}
\examples{
mod <- load_nhanes()
}
