% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mlnhanes.R
\docType{data}
\name{mlnhanes}
\alias{mlnhanes}
\title{NHANES Machine Learning Dataset}
\format{
A data frame with ~35,000 rows and ~455 variables
}
\source{
NHANES (CDC)
}
\usage{
mlnhanes
}
\description{
Prepared NHANES dataset for teaching
machine learning methods in epidemiology.
}
\keyword{datasets}
